Add a simple script to generate serialization code
This is usually performed as a cmake step, but I'm not sure we can rely
on Android Studio running the correct tools for it, so I'm just putting
this small script here. Currently it does not check for errors and
assumes everything is just fine.

You're going to need boost libraries on your host system, as well as a
g++ compiler. A good rule of thumb is, if you can build OpenApoc for
your host, you're good to go to build OpenApoc for Android.